+++ src/main/java/de/kleemann/co2_hsharz/core/transport/fuel/FuelService.java
@@ -1,13 +1,14 @@
-package de.kleemann.co2_hsharz.core.fuel;
+package de.kleemann.co2_hsharz.core.transport.fuel;
 
 import java.util.List;
 
 import org.springframework.stereotype.Service;
 
-import de.kleemann.co2_hsharz.persistence.fuel.FuelPersistanceService;
+import de.kleemann.co2_hsharz.persistence.transport.fuel.FuelPersistanceService;
+import de.kleemann.co2_hsharz.persistence.transport.fuel.TransportMediumFuel;
 
 /**
- * This Service provides...
+ * This Service provides functionality to get all {@link TransportMediumFuel} mapped to {@link FuelImpl}
  * 
  * @author Fabian Siemens
  * @version 1.0
@@ -21,6 +22,10 @@ public class FuelService {
 		this.fuelPersistanceService = fuelPersistanceService;
 	}
 	
+	/**
+	 * Find all available {@link TransportMediumFuel} and mappes them to a list of {@link FuelImpl}
+	 * @return {@link List} of {@link FuelImpl}
+	 */
 	public List<FuelImpl> findAllFuels(){
 		return fuelPersistanceService.findAllFuels()
 				.stream()
